Refactor into three scripts
* testsystem steup
* on-the-fly noneq adaptation
* passing the optimized protocol to timemachine for "deployment"
